requirements.txt
Flask
Pillow
gunicorn
cloudinary
requests
